Support reading DICOMs directly from tarballs
The DICOM template string can match multiple tarballs for one subjects,
in which case the filegroup keys get prefixed with the tarball index.
Tarballs are sorted, so that chronological naming schemes result in
chronological prefixes.

This simplifies things of mutliple scan sessions are performed in
immediate succession and the included runs shall be treated as coming
from a single session.

This change should not affect keys or behavior in the case of plain
DICOM files globbed from a directory.