RF(possibly BK): removing session embedding into seqinfo, preparing for getting arb set of files
Decided to split away 'session' notion, which was pretty much used in the script
only whenever multiple tarballs are provided, as a sign of multiple sessions.
Otherwise -- session is given on cmdline, and thus processing that particular session
might be 'incompatible' with future runs for other sessions if we save the entire
mapping in a single file which we might load, and which wouldn't have that session
information.

Also, to allow for consumption of arbitrary set of dicoms, which might be coming
from different studies and sessions, we need to analyze/group before extracting
useful session sequence information.  So, that was also in preparation to that