BF: reincarnate logging
Many messages just need to be downgraded to DEBUG level and additional
care should be taken about logging configuration for other modules we use.
But in general -- we do need logging and should be able to control it from
the environment variables for e.g. tuning it up quickly during test runs