RF(BF): move instead of copy converted files
Original motivation is that somehow prior commits trying to split apart prefix
from the full path we are already providing broke some tests, e.g. resulted in
extra commits on rewrites.

Since they could be one the same file system and otherwise not used after being
transfered to the target destination, it would be much cheaper to just move
them instead of copying.  It is also more likely to have sufficient space in
destination than in the /tmp, which already brought questions: this commit would
Closes #481 as a result as well.