ENH: anonymize_sid - do sanity check that anonymization script produced output
So it will be for the script to just return original sid if it decides to
not anonymize it, which is to some degree change in behavior.  But I think
it would be beneficial in the long run due to more straight-forward and robust
operation.  Docstring for anon_cmd does not mention that empty output implies
no change in subject id