Do not issue warning if cannot parse _task entity
Also adjusted docstring to note that now _task- can be used in many other modalities

    ❯ git grep -l -B4 -E 'task: (r|o)'
    src/schema/README.md
    src/schema/rules/files/raw/anat.yaml
    src/schema/rules/files/raw/beh.yaml
    src/schema/rules/files/raw/channels.yaml
    src/schema/rules/files/raw/eeg.yaml
    src/schema/rules/files/raw/func.yaml
    src/schema/rules/files/raw/ieeg.yaml
    src/schema/rules/files/raw/meg.yaml
    src/schema/rules/files/raw/nirs.yaml
    src/schema/rules/files/raw/pet.yaml
    src/schema/rules/files/raw/task.yaml

and overall -- even if required for some modality (e.g. func) IMHO we should
not redo bids-validator functionality here.  It should be validator to report
oddities (may be we should bolt on integration with basic validation present in
bidsschematools).