Remove "v" prefix from the tag for docker hub
And also tag as unstable as that is the one we promoted
a little in CHANGELOG in the times of 0.5 releases.

We remove "v" since not semver and IMHO not needed, and that
is how we set it up on docker hub originally so prior
releases are all without "v" and only the ones we uploaded
from actions were with "v".