RF: Gather/use list of all dicom files to avoid calling wrapper_from_file at module level
I think it is safer to avoid 3rd party library calls at the module import time
since if they are buggy it would be too loud of kaboom. Also it adds CPU time
at import time even if not used later on. So replaced with defining a list
of all DICOMs and then just using that function directly from filename.

This also would run it against more DICOMs as this time we would glob recursively.